remove parent factory
