+++ packages/better-php-doc-parser/src/ValueObjectFactory/PhpDocNode/Doctrine/JoinColumnTagValueNodeFactory.php
@@ -4,11 +4,30 @@ declare(strict_types=1);
 
 namespace Rector\BetterPhpDocParser\ValueObjectFactory\PhpDocNode\Doctrine;
 
+use Rector\BetterPhpDocParser\Printer\ArrayPartPhpDocTagPrinter;
+use Rector\BetterPhpDocParser\Printer\TagValueNodePrinter;
 use Rector\BetterPhpDocParser\ValueObject\PhpDocNode\Doctrine\Property_\JoinColumnTagValueNode;
-use Rector\BetterPhpDocParser\ValueObjectFactory\PhpDocNode\AbstractTagValueNodeFactory;
 
-final class JoinColumnTagValueNodeFactory extends AbstractTagValueNodeFactory
+final class JoinColumnTagValueNodeFactory
 {
+    /**
+     * @var ArrayPartPhpDocTagPrinter
+     */
+    private $arrayPartPhpDocTagPrinter;
+
+    /**
+     * @var TagValueNodePrinter
+     */
+    private $tagValueNodePrinter;
+
+    public function __construct(
+        ArrayPartPhpDocTagPrinter $arrayPartPhpDocTagPrinter,
+        TagValueNodePrinter $tagValueNodePrinter
+    ) {
+        $this->arrayPartPhpDocTagPrinter = $arrayPartPhpDocTagPrinter;
+        $this->tagValueNodePrinter = $tagValueNodePrinter;
+    }
+
     /**
      * @param array<string, mixed> $items
      */
