+++ packages/better-php-doc-parser/src/ValueObjectFactory/PhpDocNode/Symfony/SymfonyRouteTagValueNodeFactory.php
@@ -4,11 +4,30 @@ declare(strict_types=1);
 
 namespace Rector\BetterPhpDocParser\ValueObjectFactory\PhpDocNode\Symfony;
 
+use Rector\BetterPhpDocParser\Printer\ArrayPartPhpDocTagPrinter;
+use Rector\BetterPhpDocParser\Printer\TagValueNodePrinter;
 use Rector\BetterPhpDocParser\ValueObject\PhpDocNode\Symfony\SymfonyRouteTagValueNode;
-use Rector\BetterPhpDocParser\ValueObjectFactory\PhpDocNode\AbstractTagValueNodeFactory;
 
-final class SymfonyRouteTagValueNodeFactory extends AbstractTagValueNodeFactory
+final class SymfonyRouteTagValueNodeFactory
 {
+    /**
+     * @var ArrayPartPhpDocTagPrinter
+     */
+    private $arrayPartPhpDocTagPrinter;
+
+    /**
+     * @var TagValueNodePrinter
+     */
+    private $tagValueNodePrinter;
+
+    public function __construct(
+        ArrayPartPhpDocTagPrinter $arrayPartPhpDocTagPrinter,
+        TagValueNodePrinter $tagValueNodePrinter
+    ) {
+        $this->arrayPartPhpDocTagPrinter = $arrayPartPhpDocTagPrinter;
+        $this->tagValueNodePrinter = $tagValueNodePrinter;
+    }
+
     public function create(): SymfonyRouteTagValueNode
     {
         return $this->createFromItems([]);
